code/FontTool/requirements.txt
Pillow==4.0.*
